Add files via upload
Intro to Gonorrhea and Chlamydia is a pamphlet that can be used as a visual aid when explaining gonorrhea and chlamydia to youths; Intro pamphlet is a guide to illustrate how pay-it-forward works and can also be used as a visual aid when the implementer is introducing the pay-it-forward service delivery to youths; sampling instruction contains information to help guide youth who would like to collect rectal sample for gonorrhea and chlamydia testing by themselves. 
Those above documents are toolkit that have been tested and well received in two research studies among men who have sex with men at risk of sexually transmitted infections. Necessary adaption and translation is recommended when used in a different setting or among different populations.